requirements.txt
fastapi
uvicorn[standard]
llama-cpp-python
